requirements.txt
Flask==1.0.2
flask-cors==3.0.7
Flask-Dance[sqla]==1.2.0
flask-httpauth==3.2.4
flask-login==0.4.1
flask-mail==0.9.1
flask-migrate==2.3.1
flask-moment==0.6.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
passlib[argon2]==1.7.1
psycopg2-binary==2.7.7
pyjwt==1.7.0
pytz==2018.9
requests==2.19.1
rq==0.13.0
timezonefinder==4.0.1
uwsgi==2.0.17.1
WTForms==2.2.1
